pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "heysol-api-client"
version = "1.2.0"
description = "A comprehensive Python client for the HeySol API with MCP protocol support, built by Dexter Hadley, MD/PhD"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Dexter Hadley, MD/PhD", email = "iDrDex@HadleyLab.org"},
]
maintainers = [
    {name = "HadleyLab", email = "iDrDex@HadleyLab.org"},
]
keywords = [
    "heysol",
    "api-client",
    "mcp",
    "memory-management",
    "healthcare",
    "clinical-trials",
    "oncology",
    "python",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Internet :: WWW/HTTP",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.25.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=0.19.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=1.0.0",
    "flake8>=4.0.0",
    "pre-commit>=2.17.0",
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/HadleyLab/heysol-api-client"
Documentation = "https://docs.heysol.ai/api-reference"
Repository = "https://github.com/HadleyLab/heysol-api-client"
Issues = "https://github.com/HadleyLab/heysol-api-client/issues"
Changelog = "https://github.com/HadleyLab/heysol-api-client/blob/main/CHANGELOG.md"

[project.scripts]
heysol-client = "cli:main"

[tool.setuptools]
packages = {find = {}}
include-package-data = true

[tool.setuptools.package-data]
heysol = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.9"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_untyped_decorators = false
no_implicit_optional = false
warn_redundant_casts = false
warn_unused_ignores = false
warn_no_return = false
warn_unreachable = false
strict_equality = false

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
    "--cov=heysol",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
    "--cov-fail-under=80",
]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "async: Async tests",
]